+++ sdk/data/aztables/client_test.go
@@ -158,13 +158,14 @@ func TestMergeEntity(t *testing.T) {
 			reMarshalled, err := json.Marshal(mapEntity)
 			require.NoError(t, err)
 
-			_, updateErr := client.UpdateEntity(ctx, reMarshalled, &UpdateEntityOptions{UpdateMode: MergeEntity})
+			_, updateErr := client.UpdateEntity(ctx, reMarshalled, &UpdateEntityOptions{UpdateMode: EntityUpdateModeMerge})
 			require.Nil(t, updateErr)
 
-			var qResp ListEntitiesPage
+			var qResp ListEntitiesPageResponse
 			pager := client.List(listOptions)
-			for pager.NextPage(ctx) {
-				qResp = pager.PageResponse()
+			for pager.More() {
+				qResp, err = pager.NextPage(ctx)
+				require.NoError(t, err)
 			}
 			postMerge := qResp.Entities[0]
 			var unmarshalledPostMerge map[string]interface{}
@@ -190,7 +191,7 @@ func TestMergeEntityDoesNotExist(t *testing.T) {
 			marshalled, err := json.Marshal(entityToCreate)
 			require.NoError(t, err)
 
-			_, updateErr := client.UpdateEntity(ctx, marshalled, &UpdateEntityOptions{UpdateMode: MergeEntity})
+			_, updateErr := client.UpdateEntity(ctx, marshalled, &UpdateEntityOptions{UpdateMode: EntityUpdateModeMerge})
 			require.Error(t, updateErr)
 		})
 	}
@@ -207,7 +208,7 @@ func TestInsertEntity(t *testing.T) {
 			marshalled, err := json.Marshal(entityToCreate)
 			require.NoError(t, err)
 
-			_, err = client.InsertEntity(ctx, marshalled, &InsertEntityOptions{UpdateMode: ReplaceEntity})
+			_, err = client.InsertEntity(ctx, marshalled, &InsertEntityOptions{UpdateMode: EntityUpdateModeReplace})
 			require.NoError(t, err)
 
 			filter := "RowKey eq '1'"
@@ -229,14 +230,15 @@ func TestInsertEntity(t *testing.T) {
 			require.NoError(t, err)
 
 			// 4. Replace Entity with "bool"-less entity
-			_, err = client.InsertEntity(ctx, reMarshalled, &InsertEntityOptions{UpdateMode: ReplaceEntity})
+			_, err = client.InsertEntity(ctx, reMarshalled, &InsertEntityOptions{UpdateMode: EntityUpdateModeReplace})
 			require.Nil(t, err)
 
 			// 5. Query for new entity
-			var qResp ListEntitiesPage
+			var qResp ListEntitiesPageResponse
 			pager := client.List(list)
-			for pager.NextPage(ctx) {
-				qResp = pager.PageResponse()
+			for pager.More() {
+				qResp, err = pager.NextPage(ctx)
+				require.NoError(t, err)
 			}
 			postMerge := qResp.Entities[0]
 			var unmarshalledPostMerge map[string]interface{}
@@ -263,10 +265,10 @@ func TestInsertEntityTwice(t *testing.T) {
 			marshalled, err := json.Marshal(entityToCreate)
 			require.NoError(t, err)
 
-			_, err = client.InsertEntity(ctx, marshalled, &InsertEntityOptions{UpdateMode: ReplaceEntity})
+			_, err = client.InsertEntity(ctx, marshalled, &InsertEntityOptions{UpdateMode: EntityUpdateModeReplace})
 			require.NoError(t, err)
 
-			_, err = client.InsertEntity(ctx, marshalled, &InsertEntityOptions{UpdateMode: ReplaceEntity})
+			_, err = client.InsertEntity(ctx, marshalled, &InsertEntityOptions{UpdateMode: EntityUpdateModeReplace})
 			require.NoError(t, err)
 		})
 	}
@@ -291,10 +293,11 @@ func TestQuerySimpleEntity(t *testing.T) {
 			list := &ListEntitiesOptions{Filter: &filter}
 			expectedCount := 4
 
-			var resp ListEntitiesPage
+			var resp ListEntitiesPageResponse
 			pager := client.List(list)
-			for pager.NextPage(ctx) {
-				resp = pager.PageResponse()
+			for pager.More() {
+				resp, err := pager.NextPage(ctx)
+				require.NoError(t, err)
 				require.Equal(t, len(resp.Entities), expectedCount)
 			}
 
@@ -331,7 +334,7 @@ func TestQueryComplexEntity(t *testing.T) {
 
 			// Add 5 entities
 			entitiesToCreate := createComplexEntities(5, "partition")
-			for _, e := range *entitiesToCreate {
+			for _, e := range entitiesToCreate {
 				marshalledEntity, err := json.Marshal(e)
 				require.NoError(t, err)
 				_, err = client.AddEntity(ctx, marshalledEntity, nil)
@@ -342,10 +345,10 @@ func TestQueryComplexEntity(t *testing.T) {
 			expectedCount := 4
 			options := &ListEntitiesOptions{Filter: &filter}
 
-			var resp ListEntitiesPage
 			pager := client.List(options)
-			for pager.NextPage(ctx) {
-				resp = pager.PageResponse()
+			for pager.More() {
+				resp, err := pager.NextPage(ctx)
+				require.NoError(t, err)
 				require.Equal(t, expectedCount, len(resp.Entities))
 
 				for idx, entity := range resp.Entities {
@@ -354,13 +357,13 @@ func TestQueryComplexEntity(t *testing.T) {
 					require.NoError(t, err)
 
 					require.Equal(t, model.PartitionKey, "partition")
-					require.Equal(t, model.RowKey, (*entitiesToCreate)[idx].RowKey)
-					require.Equal(t, model.Integer, (*entitiesToCreate)[idx].Integer)
-					require.Equal(t, model.String, (*entitiesToCreate)[idx].String)
-					require.Equal(t, model.Bool, (*entitiesToCreate)[idx].Bool)
-					require.Equal(t, model.Float, (*entitiesToCreate)[idx].Float)
-					require.Equal(t, model.DateTime, (*entitiesToCreate)[idx].DateTime)
-					require.Equal(t, model.Byte, (*entitiesToCreate)[idx].Byte)
+					require.Equal(t, model.RowKey, (entitiesToCreate)[idx].RowKey)
+					require.Equal(t, model.Integer, (entitiesToCreate)[idx].Integer)
+					require.Equal(t, model.String, (entitiesToCreate)[idx].String)
+					require.Equal(t, model.Bool, (entitiesToCreate)[idx].Bool)
+					require.Equal(t, model.Float, (entitiesToCreate)[idx].Float)
+					require.Equal(t, model.DateTime, (entitiesToCreate)[idx].DateTime)
+					require.Equal(t, model.Byte, (entitiesToCreate)[idx].Byte)
 				}
 			}
 		})
@@ -398,29 +401,32 @@ func TestContinuationTokens(t *testing.T) {
 			require.NoError(t, err)
 
 			pager := client.List(&ListEntitiesOptions{Top: to.Int32Ptr(1)})
-			var pkContToken *string
-			var rkContToken *string
-			for pager.NextPage(ctx) {
-				require.Equal(t, 1, len(pager.PageResponse().Entities))
-				pkContToken = pager.NextPagePartitionKey()
-				rkContToken = pager.NextPageRowKey()
+			var pkContToken string
+			var rkContToken string
+			for pager.More() {
+				resp, err := pager.NextPage(ctx)
+				require.NoError(t, err)
+				require.Equal(t, 1, len(resp.Entities))
+				require.NotNil(t, resp.ContinuationNextPartitionKey)
+				require.NotNil(t, resp.ContinuationNextRowKey)
+				pkContToken = *resp.ContinuationNextPartitionKey
+				rkContToken = *resp.ContinuationNextRowKey
 				break
 			}
 
-			require.NoError(t, pager.Err())
 			require.NotNil(t, pkContToken)
 			require.NotNil(t, rkContToken)
 
 			newPager := client.List(&ListEntitiesOptions{
-				PartitionKey: pkContToken,
-				RowKey:       rkContToken,
+				PartitionKey: &pkContToken,
+				RowKey:       &rkContToken,
 			})
 			count := 0
-			for newPager.NextPage(ctx) {
-				count += len(newPager.PageResponse().Entities)
+			for newPager.More() {
+				resp, err := newPager.NextPage(ctx)
+				require.NoError(t, err)
+				count += len(resp.Entities)
 			}
-
-			require.NoError(t, pager.Err())
 			require.Equal(t, 9, count)
 		})
 	}
@@ -439,30 +445,33 @@ func TestContinuationTokensFilters(t *testing.T) {
 				Top:    to.Int32Ptr(1),
 				Filter: to.StringPtr("Value le 5"),
 			})
-			var pkContToken *string
-			var rkContToken *string
-			for pager.NextPage(ctx) {
-				require.Equal(t, 1, len(pager.PageResponse().Entities))
-				pkContToken = pager.NextPagePartitionKey()
-				rkContToken = pager.NextPageRowKey()
+			var pkContToken string
+			var rkContToken string
+			for pager.More() {
+				resp, err := pager.NextPage(ctx)
+				require.NoError(t, err)
+				require.Equal(t, 1, len(resp.Entities))
+				require.NotNil(t, resp.ContinuationNextPartitionKey)
+				require.NotNil(t, resp.ContinuationNextRowKey)
+				pkContToken = *resp.ContinuationNextPartitionKey
+				rkContToken = *resp.ContinuationNextRowKey
 				break
 			}
 
-			require.NoError(t, pager.Err())
 			require.NotNil(t, pkContToken)
 			require.NotNil(t, rkContToken)
 
 			newPager := client.List(&ListEntitiesOptions{
-				PartitionKey: pkContToken,
-				RowKey:       rkContToken,
+				PartitionKey: &pkContToken,
+				RowKey:       &rkContToken,
 				Filter:       to.StringPtr("Value le 5"),
 			})
 			count := 0
-			for newPager.NextPage(ctx) {
-				count += len(newPager.PageResponse().Entities)
+			for newPager.More() {
+				resp, err := newPager.NextPage(ctx)
+				require.NoError(t, err)
+				count += len(resp.Entities)
 			}
-
-			require.NoError(t, pager.Err())
 			require.Equal(t, 4, count)
 		})
 	}
@@ -527,9 +536,10 @@ func TestAzurite(t *testing.T) {
 
 	count := 0
 	pager := client.List(nil)
-	for pager.NextPage(ctx) {
-		count += len(pager.PageResponse().Entities)
+	for pager.More() {
+		resp, err := pager.NextPage(ctx)
+		require.NoError(t, err)
+		count += len(resp.Entities)
 	}
-	require.NoError(t, pager.Err())
 	require.Equal(t, 1, count)
 }